chore: renaming PagedResultClass to QueryableExtentions and improve styling
