+++ TcgPocket/Features/Cards/Queries/GetAllCardsPaginatedQuery.cs
@@ -5,7 +5,7 @@ using TcgPocket.Data;
 using TcgPocket.Shared;
 using TcgPocket.Shared.Dtos;
 using TcgPocket.Shared.PagedResult;
-using static TcgPocket.Shared.PagedResult.PagedResultClass;
+using static TcgPocket.Shared.PagedResult.QueryableExtensions;
 
 namespace TcgPocket.Features.Cards.Queries;
